collegelist component added
